add 1.9 extensions
