+++ src/FiniteDiff.jl
@@ -38,6 +38,16 @@ include("gradients.jl")
 include("jacobians.jl")
 include("hessians.jl")
 
-
+if !isdefined(Base,:get_extension)
+    using Requires
+    function __init__()
+        @require BandedMatrices="aae01518-5342-5314-be14-df237901396f" begin
+            include("../ext/FiniteDiffBandedMatricesExt.jl")
+        end
+        @require BlockBandedMatrices="ffab5731-97b5-5995-9138-79e8c1846df0" begin
+            include("../ext/FiniteDiffBlockBandedMatricesExt.jl")
+        end 
+    end
+end
 
 end # module
